src/sparkbaas-cli/requirements.txt
rich>=13.0.0
pyfiglet>=0.8.0
argparse>=1.4.0
docker-compose>=1.29.2
pyyaml>=6.0
python-dotenv>=1.0.0
questionary>=2.0.0
watchdog>=3.0.0
tabulate>=0.9.0
tqdm>=4.66.0
cryptography>=40.0.0